moved ui components to their own folder.
