+++ src/components/Layout.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Navbar from './Navbar';
+import Navbar from './ui/Navbar';
 
 const Layout = ({children}) => {
     return (
